US17 Test Document
System testing for US17 is done,  please review, sorry for the delay in upload, was trying to figure out how to automate the workflow before then and realize that it was not possible, hence testing came in late. 

Please review US17 and fill up the developer response under the remarks column.