US22 Test Document
US22 - System testing for US22  is done, please review the test result, another round of testing will be done as there are failing due to the developer's code did not have full coverage of the test case. Since option "2", the US3 Load save game is not implemented as of Sprint 4. Moreover, the developer code 
did not have full coverage of the test cases as invalid input entered in the main menu will cause the program to run into error